+++ src/lib/builders/context-menu/create.ts
@@ -1,4 +1,4 @@
-import { usePopper } from '$lib/internal/actions/index.js';
+import { usePopper, type InteractOutsideEvent } from '$lib/internal/actions/index.js';
 import {
 	FIRST_LAST_KEYS,
 	addMeltEventListener,
@@ -11,7 +11,6 @@ import {
 	getPortalDestination,
 	getPreviousFocusable,
 	isHTMLElement,
-	isLeftClick,
 	kbd,
 	noop,
 	omit,
@@ -20,6 +19,7 @@ import {
 	toWritableStores,
 	withGet,
 	type WithGet,
+	portalAttr,
 } from '$lib/internal/helpers/index.js';
 import type { MeltActionReturn } from '$lib/internal/types.js';
 import type { VirtualElement } from '@floating-ui/core';
@@ -113,19 +113,20 @@ export function createContextMenu(props?: CreateContextMenuProps) {
 	);
 	const longPressTimer = withGet.writable(0);
 
-	function handleClickOutside(e: PointerEvent) {
+	function handleClickOutside(e: InteractOutsideEvent) {
 		rootOptions.onOutsideClick.get()?.(e);
-		if (e.defaultPrevented) return;
+		if (e.defaultPrevented) return false;
 
 		const target = e.target;
-		if (!(target instanceof Element)) return;
+		if (!(target instanceof Element)) return false;
 
 		const isClickInsideTrigger = target.closest(`[data-id="${ids.trigger.get()}"]`) !== null;
 
 		if (!isClickInsideTrigger || isLeftClick(e)) {
-			rootOpen.set(false);
-			return;
+			return true;
 		}
+
+		return false;
 	}
 
 	const isVisible = derivedVisible({
@@ -147,7 +148,7 @@ export function createContextMenu(props?: CreateContextMenuProps) {
 				id: $menuId,
 				'aria-labelledby': $triggerId,
 				'data-state': $isVisible ? 'open' : 'closed',
-				'data-portal': $portal ? '' : undefined,
+				'data-portal': portalAttr($portal),
 				tabindex: -1,
 			} as const;
 		},
@@ -174,11 +175,14 @@ export function createContextMenu(props?: CreateContextMenuProps) {
 							open: rootOpen,
 							options: {
 								floating: $positioning,
-								clickOutside: $closeOnOutsideClick
-									? {
-											handler: handleClickOutside,
-									  }
-									: null,
+								modal: {
+									closeOnInteractOutside: $closeOnOutsideClick,
+									onClose: () => {
+										rootOpen.set(false);
+									},
+									shouldCloseOnInteractOutside: handleClickOutside,
+									open: $isVisible,
+								},
 								portal: getPortalDestination(node, $portal),
 								escapeKeydown: $closeOnEscape ? undefined : null,
 							},
@@ -339,4 +343,11 @@ export function createContextMenu(props?: CreateContextMenuProps) {
  */
 function isTouchOrPen(e: PointerEvent) {
 	return e.pointerType !== 'mouse';
+}
+
+export function isLeftClick(event: InteractOutsideEvent): boolean {
+	if ('button' in event) {
+		return event.button === 0 && event.ctrlKey === false && event.metaKey === false;
+	}
+	return true;
 }