+++ src/routes/repro/+page.ts
@@ -2,5 +2,5 @@ import { redirect } from '@sveltejs/kit';
 import type { PageLoad } from './$types.js';
 
 export const load: PageLoad = async () => {
-	redirect(303, 'https://stackblitz.com/github/melt-ui/playground');
+	redirect(303, 'https://stackblitz.com/github/melt-ui/template');
 };